fix "pass-thru" nodes #169
When a node has a moveTo in its onArrive, it works only very spuriously,
probably due to an async-y race condition.
The onArrive export is async, so this seems like the proper way to do it
anyway.
Hopefully this does not clog up / hang the project server under certain
conditions? (there is no timeout...)